Introduce CI (#6)
Introduce a CI for building and testing

The created .yml uses installing mojo and max according to the specifications of the project's .toml files. All files in the project are compiled and the tests run. In the future, the CI should also include linter and formatter, but this is not included in this PR.