Implement a BitMask graph for quick BitMask lookup (#17)
The graph facilitates quick access of items identified by
bitmasks if an item with a somewhat similar bitmask is known.

The graph will be used for quick Archetype access. It is implemented
via a list of inline arrays and nodes are identified via their index
in this list.

During the implementation of this graph, a bug in the hash function
of SIMD was detected, which is why we circumvented this via a
naively implemented dict, which will be replaced by the Mojo dict
in due time.