Create new / unified benchmarks (#22)
Switch to the builtin benchmarking system. The limitation 
is that the output is fixed to ms. Therefore, we do not 
benchmark individual actions but a whole bunch. Clear 
documentation should facilitate interpretation of the 
results anyway. Using the default framework allows better 
presentation of the results and may in the future also 
overcome the current limitations.

--- The steps taken in this commit: ---

* Add additional test / benchmarks for world

* Add tests and benchmarks for dictionary

* Use the builtin benchmarking tools

These have somewhat nicer output and
eventually this transition was necessary anyway.

* Let the CI run on all PRs