Implement component manipulation (#23)
* Add possibility to mark elements in the array list as deletable

This is needed to reuse memory for unneeded archetypes

* Implement removal of entities

Also, fix test for world

* Disable test of dict

The test fails due to a Mojo bug.

* Apply formatter

* Implement world.has

* Add benchmarks

* Improve comments, add asserts

Also: add unsafe option for get and set in Archetype

* Remove unneeded raises, get component ID array

* Let Archetype contain mask, check existence of multiple components

* Implement add and remove of components, remove unneeded code

* Imporve initialization of archetypes by making the mask optional

* Update Mojo

This required a change in the initializer syntax, as __init__ is a
static method now. Also, the ExplicitlyCopyable trait
changed.

* Fix the world tests, which were wrong

The wrong test masked a bug that is now fixed.

* Apply formatter

* Add option to initialize Archetype based on BitMask

This is helpful if the components are not known at compile
time

* Implement the option to get the indices of the set bits

This can be helpful if the components in a bitmask should be
read. Benchmarks showed that the operation is rather slow,
primarily due to writing the results into an array. Therefore,
the new iterator is never used and may be removed in the
future.

* Rename ArrayList.append to add, return index

The added elements may not be added at the List's end
but rather overwrite existing elements.

For this reason, it is important to return the element's index
when adding.

* Make all functions in the component manager static

This may allow for compiler optimizations

* Fix that new archetypes are added by their list

Before, archetypes were not correclty identified
if they were derived from existing archetypes

* Add tests

* Make Node explicitly copyable

This allows using it in containers that require this trait

* Fix benchmarks and add new ones

The capturing closures did not work properly.
Therefore, we moved the world creation into the
measured functions. The overhead is low, but it
remains to be checked whether the issue is on the
mojo or larecs side.

* Apply formatter to graph

* Activate all world benchmarks

* Fix benchmark

* Inlining for small performance improvements

* Implement exchange

* Apply formatter, adjust benchmark to Arche benchmarks

* Fix that the promised archetypes exists when benchmarking

* Re-enable bitmask benchmarks

* Add benchmark for getting the component array, rename component benchmarks

* Prevent inlining in world benchmarks

This gives more accurate results

* Remove achetype cleanup and change inlining for better performance

* Adjust tests to disabled archetype removal

* Apply formatter

* Adjust Exchange benchmark to Arche test case

* Adjust benchmarks to Arche for better comparison

* Small performance improvements

When initializing an archetype based on a bitmask,
constrain the for loop for filling the archetype to
the total number of components.

Use a power of 2 for the archetype capacity.

* Adjust archetype test to the recent changes

* Apply formatter

* Cheaper test for component existence based on bitmasks

Unused methods of archetype are removed.

* Constrain components to be unique