Implement resource (#66)
Implement resources for the world. These are handed over at world initialization.

Note that providing resources is a must, unless their type is explicitly specified as a parameter.

As a side, the archetype container was changed to a standard list. This reduces the code base that needs to be maintained. This was done during this commit to assert that an occurring issue was not due to the chained array list.

Known issue: capturing closures with World cause a segmentation fault. This is likely due to Mojo.

* Simplify component

* First attempt at implementing resources

* Solve the origin issue by copying code from tuple

* Add documentation

* Add componts to world

* Implement resource addition and removal

* Fix docs, fix ownership

* Support default constructor

* Make types movable and copyable

* Change the archetype container to list, refactor

The archetype container does not need to be
pointer persistent. With the list as data
structure, less code needs to be maintained in this package.
Also, the performance might be slightly better.

* Adjust the tests

* Export Resources

* Remove unneeded chained array list code

* Remove remainders of ChainedArrayList

* Adjust benchmarks and examples to use resources

* Fix docstring for doc test

* Update readme