Implement world.remove_entities and len(world)
For this, it is necessary to pass query as an argument to
methods of world. This could lead to multiple pointers accessing
the same memory, which is forbidden. Therefore, a new type
QueryInfor is introduced, which just holds a queries' masks and
can be implicitly constructed from a query.

len(world) is useful for testing.