Implement batch removal of entities (#86)
* Implement world.remove_entities and len(world)

For this, it is necessary to pass query as an argument to
methods of world. This could lead to multiple pointers accessing
the same memory, which is forbidden. Therefore, a new type
QueryInfor is introduced, which just holds a queries' masks and
can be implicitly constructed from a query.

len(world) is useful for testing.

* Refactor entity generation

Guideline is to use full words. Use "generation" instead of "gen"
for better readability.

* Add possibility to clear the archetype and read access to entity list

* Change type of FlexibleComponent (used in testing) for better representation string

* Add tests for len(world) and world.remove_entities

* Add benchmarks for batch removal

* Reduce batch size to allow benchmarks of expensive functions to terminate earlier