Flexible resources (#90)
Implement flexible resources that can be added at runtime. 
This also fixes a block for the implementation of systems.

* Remove leftover comment in world

* Update Mojo and make necessary adjustments

* Add a utility function for checking mutability of things

* Implement a type-agnostic container with vtable

* Adjust initializer syntax to updated mojo

* Change resources to use the flexible type

Now the resource types do not need to be specified at
compile time. This solves issues blocking the implementation of
systems and allows a more flexible handling of resources.

* First attempt of supporting both flexible and static resources

This does not work well in practise, because every function
taking a world would need to specify the resource type.

* Choose dynamic resources only

* Add inlines for better performance

* Implement a more rigorous system for distinguishing types

* Adjust satellites example

* Adjust docs

* Add tests for the type map and type ID