Move resources instead of copying them (#99)
* Update readme, introduce changelog

Co-authored-by: Martin Lange <44003176+mlange-42@users.noreply.github.com>

* Invoke move instead of copy when adding resources

If resources are large, this is much more performant.

Note: move is neither necessary nor possible with components,
as these are always memcopied to archetypes.

* Prevent the destruction of the moved values.

Otherwise, this may lead to UB.

* Update Mojo and make necessary adjustments

Note that infer-only parameters seem to not be
correctly deduced at the moment. The therefore
needed changes can be reverted at a later time.

* Update Mojo

* Disable address sanitizer when testing

Currently, the ASAN seems to slow down the test extremely,
which is why disabling it seems the better option for the time being.

* Add benchmarks for resources

* Make benchmarks shorter

* Fix AOS benchmarking script

Implicitly converting Mojo lists to Python lists
is currently not supported.

---------

Co-authored-by: Martin Lange <44003176+mlange-42@users.noreply.github.com>