Remove implicit conformance, switch to stable (#101)
Implicit trait conformance is deprecated. Hence, all trait conformance must be explicitly stated. As a result, traits should not be composed via inheritance but via a direct composition (alias &).

We switch to the stable branch so as to provide interoperability with other packages.

Individual commits:
---------------------

* Replace the deprecated CollectionElement and CollectionElementNew traits

* Split the IdentifiableCollectionElement trait in favour of trait composition

This is in line with the explicit trait conformance requirement

* Move to explicit trait conformance

* Introduce the composed resource type alias

* Switch to stable branch

* Update version numbers

* Update lock files

* Remove address sanitizer when testing the package

The address sanitizer currently has a very poor performance
and the tests without are already good for the job.

* Update changelog