Cosmetics and refactoring (#105)
Move from the deprecated @value decorator to @fieldwise_init along with explicit trait conformance.
Refactor ComptimeOptional to StaticOptional so as to match the stdlib's naming scheme and adjust it's API to the API of Optional.

* Return to readable literals

* Replace the deprecated @value with the new @fieldwise_init

This requires explicit specification of Copyable and Movable
if these are needed.

* Add explicit trait conformance, make Resources ExplicitlyCopyable only

* Rename ComptimeOptional and adjust API to Optional

StaticOptional is more in line with the stdlib's naming scheme.
Additional change: use StaticOptional where Optional was
used before without being needed.

* Adjust the source file name of StaticOptional for consistency

* Mention refactoring in changelog in case outsiders use ComptimeOptional

* Add explicit trait conformance in docs