Update mojo, adjust StaticOptional, remove StupidDict (#110)
* Switch to nightly, update lock file

* Update Mojo, remove stupid dict

Note, due to a bug in Mojo, structs are currently not
supported in test files. These tests have been disabled and
will be re-enabled as soon as the issue has been fixed.

* Rework StaticOptional to build on InlineArray

This reduces the fraction of unsafe code,
simplifies the code, removes biolerplate, and
even saves move operations on copy.

* Remove obsolete code

* Fix move test

* Fix missing fill argument in InlineArray init, remove unneeded imports

* Fix traits, remove boilderplate code

* Adjust required channels, update lock files

* Update modo

* Fix doc gen

* Disable doc tests temporarily (which fail due to a mojo bug)