Introduce ArchetypeListIterator to iterate over a given list of archetypes
In some operations we already know which archetypes where changed. Therefore it is more efficient to store a list of these archetypes and return an iterator that uses this list to iterate over all archetypes.
This can also be used for the _EntityIterator to also iterate through all entities of the iterated archetypes.