Fix duplication of entities during batch `world.add`
In the batched version of `world.add` an entity was added once per new component to the `new_archetype`. So if two new components were added the `new_archetype` was extended the amount of entities in `old_archetype` TIMES TWO. This should not happen as it will fill the available memory very quickly!