Implement Batch addition (#113)
* Implement batch component addition as overload of `world.add`

* Optimize `archetype.reserve` to reduce frequent reallocations

If requested Archetype capacity exceeds available increase to next sufficient power of 2

* Add StaticVariant

StaticVariant can model a value that can have multiple types but it is known at compile-time which type will be used.
It may be helpful to think of it as the Variant type from stdlib, but resolved completely at compile-time.

* Introduce ArchetypeByListIterator to iterate over a given list of archetypes

In some operations we already know which archetypes where changed. Therefore it is more efficient to store a list of these archetypes and return an iterator that uses this list to iterate over all archetypes.
This can also be used for the _EntityIterator to also iterate through all entities of the iterated archetypes.

* Add benchmarks for batch adding components to entities

* Add doc section about batch component addition

* Add function to calculate next power of 2 fast

* Add helper `QueryInfo.matches` to encapsulate query matching logic

* Add `BitMask.__or__`